Bug fixes
In detail:

1. Notify of removed udp flows *after* the MT-specific checks of
close_udp_flow(), rather than before.  When the notification was done
before, some spurious events were triggered, when running in -p mode,
for the flows that had a zeroed pup->last_time.  Note that an udp
flow has a zeroed pup->last_time for the little amount of time between
its creation - in newUTP - and its initialization - udp_flow_stat, at
udp.c line 430.  In other words, we're speaking of a race.  Putting the
notification *after* the MT checks of close_udp_flow() removes the race
hazard, because there is one check that makes sure that pup->last_time
is different from {0, 0}.

2. Make number of active flows signed.  This will make it easier to
catch logic errors that cause more REMOVE_FLOW than CREATED_FLOW events,
because (IMO) it's simpler to spot -1 than an huge positive number when
reading debugging output.  This patch had been written while tracking
down the bug explained in #1.

3. Make sure endpoint_type_to_string() buffer is *always* initialized to
'\0'.  This fixes a bug that caused random data to be printed, when a
TCP endpoint was classified as UNKNOWN.

4. Don't print anything on the standard output.  It's more coherent to
always use the same standard flow for all log messages (and we've chosen
to use stderr.)

5. Adjust MAX_ENDPOINTS value.  There is a detailed comment that explains
the reason why we've chosen such value, and I don't want to duplicate it.