Implement endpoint_prune() and stats
In detail:

* Add bytes/pkts statistic, that is collected every time a flow is removed.

* Add tot_unclassif/tot_classif statistic, that is collected every time a
flow is classified.  If the subtype is notzero, the flow is blessed classified,
otherwise unclassified (Note that the current implementation fires CLASSIFIED
and REMOVED events when a flow is removed, but this may change in future.)

* Add time statistic.  We record the time when an endpoint was created and the
time of the last REMOVED event we've seen.  The latter is needed to implement
endpoint_prune().

* Debug the flow events we see - i.e. CREATED_FLOW, REMOVED_FLOW and
CLASSIFIED_FLOW - to make the endpoint engine behavior more transparent.

* Implement endpoint_prune().  It is implemented as a loop that scans the
vector of endpoints and prunes all the ones with (epnt->active_flows == 0 &&
elapsed(epnt->last_removed, current_time) > ENDPOINT_IDLE_TIME).

* While at it, add basic MT support, protecting with a single mutex all the
possible critical section.  Note that this simple approach may lead to slower
performances, if compared to a more fine grained one.