Implement endpoint classification engine and endpoint_adjust_*_type()
In detail:

* Separate engine and interface in endpoint classification.  This should make
it easier and simpler to experiment in this field - as at the moment it is
neither clear which is the best algorithm to update the endpoint type, nor
obvious the right way to represent it.

* Avoid semantic confusion between endpoint_notify_classified_*() and
tot_classif/tot_unclassif fields of struct endpoint.  A flow is "classified"
after the flow classifier has taken a decision on its type.  The type may be
a "known" one (such as HTTP) or "unknown".  Thus, tot_classif/tot_unclassif
are actually counters of the number of classified flows whose type is
known/unknown.  For this reason, we rename them tot_known/tot_unknown.

* Implement endpoint_adjust_*_type() operation, that adjusts the type of a flow
using the flow endpoints[1] types.  Usually the type is adjusted only if the
flow type is unknown, but it is possible to force the operation to adjust it in
any case.

* Use verbose defines rather than TRUE and FALSE, whenever possible, to make the
code more informative and self-explaining.  For example, TRUE/FALSE are less
self-explaining than ENDPOINT_ADJUST_ALWAYS/ENDPOINT_ADJUST_IF_UNKNOWN.

Note
====
[1] The "flow endpoints" are the two endpoints that make a flow.