Implement endpoint_print_stats (and much much more...)
In detail:

1. Implement endpoint_print_stats() that prints statistics for each endpoint.
It is invoked when an endpoint is about to be pruned.

2. Renamed ->type_to_string() engine operation ->format_type_name().  The latter
name is more consistent with the ->format_type() operation added in #1.

3. tcp: adjust_type() before we make_conn_stats() for the tcp flow that is
about to be closed.  With the old code, the (possibly) adjusted flow type
wasn't used.

4. Uniform algorith for the two notify_classified_*() functions.  That is,
they have been rewritten to be as equal as possible.

5. Make the IMA engine more robust, by defining a NO_TYPE default type, and
adding more checks to catch the case where the type is not UDP, TCP or TCP_P2P.

6. More precise event debugging: don't say that an udp flow has been classified
when actually it is an udp direction.

7. Modify debugging timestamp format.  The new format is the raw one, which is
possible to get easily on both ethereal and tcpdump.  This should make debugging
and tracing even more easier.

8. Rewritten IMA_engine_update_type_core().  The new code is more readable and
use 33% of the lines of the old one.

9. Adjust flow type only if flow endpoint type is ASSURED.  In my opinion, there
is no point in adjusting, if we're not sure of the endpoint type.

10. Add doc/endpoint.txt documentation, that explains the design choices, and
outlines the current implementation.

11. Sparse cosmetic changes.