Add RIMA (Revised IMA) engine (and much more...)
In detail:

1. IMA: reset type and subtype in MAYBE -> INITIAL transition.  This helps
RIMA engine implementation.

2. Add RIMA (Revised IMA) engine.  This engine tries to cope with possible
IMA engine problems, such as:

i. The fact that each direction of an UDP flow may be classified
independently of the other.  It may be a problem if the two directions
ends up with different labels.

ii. The fact that we save just the higher layer tcp flow label.  It may
be a problem if we want the inherit operation to adjust *exactly* tcp
flow labels (i.e. both con_type and p2p_type.)

To workaround the former, we store in type the client to server label,
and in subtype the other direction label.  To workaround the latter, we
store in type tcp_pair con_type, and in subtype p2p_type.

3. Add ->select_best() to endpoint engine model.  It allows engines to
use their proper algorithm to choose the best flow endpoint for the adjust
operation.

4. Print debugging information only if debug is enabled (via -d command line
switch.)

5. Improved documentation.

6. Sparse cosmetics.