Implement L7hist engine
In detail:

* Implement L7hist engine, that keeps L7 types frequency histogram -- one
histogram for TCP endpoints, two (one per direction) for UDP ones.  We
implement that using a linkd list rather than a table, to consume less memory,
as we expect sparse histograms.

* Classification: Ignore data-less TCP flows and UDP half flows.  A data-less
(half)flow must be different from an unclassified (half)flow, and must not be
classified as such.

* Implement endpoint_write_logfile_header(), that writes a comment explaining 
file field semantic at the beginning of each log file.