Another batch of fixes, improvements, bells and whistles
In details:

1. Add -E tcp_idle=TIME / -E udp_idle=TIME command line options, that
trim the before prune time for tcp and udp endpoints.

2. Add 'internal' information to endpoint structure.

3. Default to L7hist engine.  It seems to be the best one.

4. In documentation and code, use the term 'flow-dir' instead of
'half-flow'.  The former seems to be more informative.

5. Switch from boolean adjust parameter to adjust *policy*.  There are
three policies:

    a) never: Never adjust flow type.

    b) if_needed: Adjust flow type only if it is unknown.

    c) always: Always adjust flow type.

The related command line option is -E adjust=POLICY.
The default policy is <never>.

6. IMA engine: When adjusting type, zero the fields we're not adjusting.

7. RIMA engine: Likewise.

8. Add endpoint_timeconv, a mechanism to make string to integer time
conversion more flexible: Now an optional suffix may be specified to
set the proper measure unit, such as hours, minutes, or seconds.

It *was* odd to have to specify idle_time in microseconds ;-).

9. Update documentation and comments.