(Back)ports
In detail:

1. Initial port to OpenBSD.  There is some more work to do.  This
patch makes it possible to compile Tstat successfully under OpenBSD
(by the way it works, but I've not tested it very well: I'm using
my laptop for smoke testing only.)

2. Improve code that ignores TCP source endpoint.  Testing, I've
found out that it is better to comment out all the places where we
try to get TCP source endpoint, or we'll foo our statistics a bit
(actually this finding was a serendipity, because I coded the new
endpoint module I'm working on with all these places commented out.)
By the way, this seems to me a bit strange and I should second-look
at it.  Perhaps it's not the case to ignore TCP source endpoints.
Should I ignore it or not, the good news its that the after-patch
behavior is more correct.

3. Backport default settings.  From the new endpoint module I'm
working on.  We can't replace the old module with the new one until
they behave (as much as possible) in the same way.

4. Fix self-sending-UDP-endpoint bug.  Comparing the old and the
new endpoint modules I've found that the old one misbehaves when
there is an UDP endpoint that is self-sending UDP datagrams -- by
the way this should not happen because the OS should route the IP
packets locally :-P.  This fix is backported from the new module.

5. Fix zombie endpoints bug.  We define zombie an idle endpoint
that has been idle for more than idle-time.  Comparing the old and
the new endpoint modules I've found that the old one may re-use
zombie endpoints in certain conditions, leading to more aggregated
results than expected.  This patch makes sure that endpoint_find()
flushes zombie endpoints to logfile.  (A considerable amount of
black magic has been spelled to make the old module behave like the
new one with respect to this bug.)