Replace old endpoint module with the new one
Briefly, the new endpoint module has a better design and a better
architecture.

(Note that this commit is not going to bring you a 100% functional and
tested new endpoint module.  There are still some pending issues to be
addressed, such as early, fast and adjust[1] classification.)

In detail:

1. The new module is clearly split into two layers: a generic endpoint and
endpoint type management layer and a glue layer.  This is good because:

	1.1 The standalone generic layer could eventually be ported to
	another sniffer, without too much effort.

	1.2 Having clean layer boundaries makes it very simple to
	implement different classification strategies.	For instance,
	once complete, the new module will support normal, fast, early
	and adjust classification.

	1.3 Unless it had been a good discovery tool that helped to
	find the right way to represent endpoint types, the engine model
	pushed complexity at the lowest level, while it seems better to
	pull it as near as Tstat as possible.

2. The new module has a very clean endpoint type model, suitable for
multiple and partially overlapped classifiers.	In detail:

	2.1 The endpoint type is defined by one or more histograms,
	whose semantic is up to the glue layer.

	2.2 Histogram is not representing type PDF anymore.  For each
	type but the unknown one there's a column, counting the number of
	type-matching flows.  Each histogram keeps also a total-flows
	counter.

	2.3 Each time a classifier has classified a flow, we increment
	total-flows counter and, unless the flow is classified as unknown,
	the relevant column.

	2.4 The best histogram type is the column that has the best
	matches over total-flows ratio, that we call "type affinity."

	2.5 Multiple histograms may be merged in a single one, using
	module APIs, to calculate aggregated type affinity.

	2.6 If (aggregated) type affinity is 0% for each type, we say
	that the endpoint type is unknown.

We use flow endpoints[2] aggregated type affinity to implement fast and
adjust classification.  The type affinity is a measure of how much we
trust the best aggregated type.

3. A better cache design, using linked lists rather than a vector.
This leads to cleaner and simpler code (in particular with respect to
critical sections management.)  Furthermore, such new design has always
been faster than the old one, being 7.5% faster in one case and 1.8%
in the other.  While speed was not a goal, it's good to notice that
we've not made it slower.

4. An allocator that may either use Tstat-style linked lists or malloc()
and free().  This exists to make the module more portable.

5. A functional programming API to iterate over flow endpoints, that
is more secure and elegant than the old notify() API.

6. Some bug fixes, that had already been backported to the old module
in previous commits.


Note
====

1. Some definitions:

- Early classification:  When you push the endpoint module flow type
before flow has terminated.  This may lead to pushing flow type more
than once.

- Fast classification:	When you classify a flow using information in
endpoint cache (if available), thus bypassing DPI analysis.

- Adjust classification:  When you adjust flow type using information
in endpoint cache, after the cache has been updated with DPI results.

2. We define flow endpoints the two endpoints that make a flow.