Improve fast and adjust algorithms
In detail:

1. Try to fast classify a flow in the create event handler rather
than in the plugin.  This is the cleanest way to do fast classify,
because it avoids conflicts between the fast classifier and the
DPI.  Thanks to Alessandro for the idea.

2. Rationalize flag names (e.g. use <OBJECT>_FLAG_<FLAGNAME> for
all flags).

3. Simplify adjust_type code path, for various operations.  Always
invoke the `simple' hook from the `adjust_type' hook rather than
duplicating the code.

4. Log_epnt_flow: Add bytes stats.  We need that to measure adjust
algorithm goodness by bytes rather than by flows.

5. Improve adjust algorithm for UDP: Use two adjust_type structures,
one per flow direction, rather than just one.  This is the cleanest
way to do adjust, because it correctly separates the two UDP flow
directions.