Bugfixes and cleanups after code audit
In detail:

1. Various random cosmetic-only changes.

2. epnt_cache.c: Use EPNT_NITEMS() macro rather than testing for
EPNT_CACHE_SIZE.  This makes the code more reusable.

3. Add some more diagnostic checks.

4. epnt_remove_tcp: Do at.total > 0 test for simmetry with other
places where we already do the same.  The check may seem useless
at the first glance, but it's only because our current threshold
is zero.  We may raise it to a more significant percentage in the
future.

5. [BUG] epnt_glue.c: When doing adjust, we must update_at() after
we've propagated flow type (rather than before) otherwise current
flow type won't be used in the merge.